refactor(quote form): simplifies rendering of quote request form
