+++ app/components/MachinesPage.tsx
@@ -1,12 +1,13 @@
 "use client";
 
 //LIBRARY IMPORTS
+import React, { useState } from "react";
 import { Grid } from "@chakra-ui/react";
 import { useTina } from "tinacms/dist/react";
 import { PageQuery } from "@/tina/__generated__/types";
 
 //LOCAL IMPORTS
-import QuoteForm from "./QuoteForm";
+import MachinesLayoutQuoteForm from "./quoteFormLayouts/MachinesLayout";
 import Machines from "./Machines";
 
 export default function MachinesPage(props: {
@@ -16,7 +17,13 @@ export default function MachinesPage(props: {
   };
   query: string;
 }): JSX.Element {
+  // SET STATES
+  const [submitSuccessful, setSubmitSuccessful] = useState(false);
+
   // HOOKS
+  const handleCloseSuccess = () => {
+    setSubmitSuccessful(false);
+  };
   const { data } = useTina(props);
 
   return (
@@ -27,7 +34,13 @@ export default function MachinesPage(props: {
             return <Machines key={i} {...block} path="/view-machines" />;
           }
           case "PageViewMachinesBlocksQuoteSection": {
-            return <QuoteForm key={i} {...block} layoutType="machinesPage" />;
+            return (
+              <MachinesLayoutQuoteForm
+                key={i}
+                {...block}
+                setSubmitSuccessful={setSubmitSuccessful}
+              />
+            );
           }
         }
       })}
