+++ content/pages/view_machines.md
@@ -6,7 +6,11 @@ blocks:
       operators who know how to get the most out of them.
     machineCards: content/machines/machines.md
     _template: machines
-  - _template: quoteSection
+  - requestQuoteForm:
+      formTitle: Request a Quote
+      field1Placeholder: Name
+      field2Placeholder: Industry Type
+    _template: quoteSection
 _template: viewMachines
 ---
